windows_requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
audioop-lts==0.2.1
audioread==3.0.1
av==14.4.0
azure-cognitiveservices-speech==1.44.0
azure-core==1.34.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
curl_cffi==0.11.1
decorator==5.2.1
distro==1.9.0
elevenlabs==0.2.27
emoji==2.14.1
enum34==1.1.10
executing==2.2.0
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
frozendict==2.4.6
fsspec==2025.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
griffe==1.7.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
halo==0.0.31
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
keyboard==0.13.5
lazy_loader==0.4
librosa==0.11.0
linkify-it-py==2.0.3
llvmlite==0.44.0
log-symbols==0.0.14
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.9.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
mss==10.0.0
multitasking==0.0.11
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
openai==1.81.0
openai-agents==0.0.16
opencv-python==4.11.0.86
openwakeword==0.6.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pvporcupine==1.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.1
pynput==1.8.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.2
PyYAML==6.0.2
realtimestt==0.3.104
RealTimeTTS==0.1.3
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spinners==0.0.24
sse-starlette==2.3.5
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
stanza==1.10.1
starlette==0.46.2
stream2sentence==0.3.1
sympy==1.14.0
termcolor==3.1.0
textual==3.2.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==15.0.1
yfinance==0.2.61
google_auth_oauthlib==1.2.2
SpeechRecognition==3.14.3